requirements.txt
Django>=4.2,<5.0
psycopg2-binary>=2.9
gunicorn>=21.2.0
python-dotenv>=1.0.0
django-debug-toolbar>=4.2.0
djangorestframework>=3.14.0
djangorestframework-simplejwt>=5.3.1
drf-yasg>=1.21.7
pandas>=2.0.0
django-celery-beat>=2.5.0
celery>=5.3.0
redis>=5.0.0
django-redis>=5.4.0
croniter>=1.4.0
openpyxl>=3.1.0
odfpy>=1.4.1
django-filter>=23.0